remove unused named exports
