+++ lib/shower/container.js
@@ -1,4 +1,4 @@
-import { EventEmitter } from '../emitter';
+import EventEmitter from '../emitter';
 import { bound } from '../utils';
 
 const getTransformScale = () => {
