+++ lib/shower/player.js
@@ -1,5 +1,5 @@
-import { EventEmitter } from '../emitter';
 import { bound, isInteractiveEl } from '../utils';
+import EventEmitter from '../emitter';
 
 /**
  * @class
