+++ lib/shower/shower.js
@@ -1,5 +1,5 @@
-import { EventEmitter } from '../emitter';
 import Options from '../options';
+import EventEmitter from '../emitter';
 import defaultOptions from './default-options';
 import Player from './player';
 import Container from './container';
