+++ lib/slide/index.js
@@ -1,5 +1,5 @@
-import { EventEmitter } from '../emitter';
 import Options from '../options';
+import EventEmitter from '../emitter';
 import { Store } from '../utils';
 import Layout from './layout';
