+++ lib/slide/layout.js
@@ -1,5 +1,5 @@
 import { bound } from '../utils';
-import { EventEmitter } from '../emitter';
+import EventEmitter from '../emitter';
 import { defaultOptions } from '../shower';
 import Options from '../options';
